[implementation] style: clean up signedness warnings (2/2)

+++ implementation/epuck/src/moderator.h
@@ -6,8 +6,8 @@
 
 typedef struct ModInputs {
     T2TData_Moderate* t2t_data;
-    int own_victim_x;
-    int own_victim_y;
+    double own_victim_x;
+    double own_victim_y;
     unsigned int found_victim_xy;
     unsigned int give_up;
 } ModInputs;
